requirements.txt
altair==5.5.0
attrs==25.4.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.2
captcha==0.7.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.3
exceptiongroup==1.3.1
extra-streamlit-components==0.1.81
gitdb==4.0.12
GitPython==3.1.45
google_search_results==2.4.2
h11==0.16.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.13.0
numpy==2.0.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.33.2
pyarrow==21.0.0
pycparser==2.23
pydeck==0.9.1
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
selenium==4.36.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
streamlit==1.50.0
streamlit-authenticator==0.4.2
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.2.0
